+++ JiraPS/Public/Get-JiraResponseHeaderLogConfiguration.ps1
@@ -1,0 +1,7 @@
+﻿function Get-JiraResponseHeaderLogConfiguration {
+    # .ExternalHelp ..\JiraPS-help.xml
+    [CmdletBinding()]
+    param()
+
+    $script:JiraResponseHeaderLogConfiguration
+}